+++ example/android/app/src/main/kotlin/com/dstk/video_compressor_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.dstk.video_compressor_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}