+++ lib/light_compressor.dart
@@ -1,0 +1,2 @@
+export 'src/compression_result.dart';
+export 'src/light_compressor.dart';